Correctly detect if the client is unicode or not and set translation config correctly at p4api level
Detection:
P4CHARSET=none means the client is not unicode. See
http://www.perforce.com/perforce/doc.current/manuals/cmdref/P4CHARSET.html

For Perforce services operating in the default (non-Unicode mode),
P4CHARSET must be left unset <b>(or set to none)</b> on user
workstations. If P4CHARSET is set, but the service is not operating in
internationalized mode, the service returns the following error message:
Unicode clients require a unicode enabled server.

(emphasis mine)

For telling p4api not to unse any translation for non-unicode servers
see http://stackoverflow.com/a/5250949/4628203 for the line
getClientApi()->SetTrans(...)